requirements.txt
aws-lambda-powertools
boto3
pyyaml